qtbase: Cherry-pick ioctl change from Qt 5.4
The log is spammed with:

May 03 06:24:58 kernel[221]: Loading kernel module for a netw....
May 03 06:24:59 kernel: Loading kernel module for a network ...d.

this is because of an ioctl for an invalid device which leads
to the kernel trying to load a module. This occurs many times
a minute.

I have cherry-picked 1910454fe00cce8b815b1abc0a18a04d3d387ccf which
merged without conflicts. It has not been compile tested.

Related: SYS#1610